+++ src/ade/__init__.py
@@ -1,6 +1,6 @@
 """Ade Programming Language & Language-Building Platform.
 
-Version: 0.3.0
+Version: 0.4.0
 """
 
-__version__ = "0.3.0"
+__version__ = "0.4.0"